requirements.txt
appdirs==1.4.4
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
build==1.2.2.post1
buildozer==1.5.0
certifi==2020.6.20
chardet==4.0.0
click==8.2.0
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
Cython==0.29.37
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flask-cors==5.0.1
fonttools==4.58.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
packaging==25.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyjnius==1.6.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproject_hooks==1.2.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
sh==1.14.3
six==1.16.0
systemd-python==234
toml==0.10.2
tomli==2.2.1
Twisted==22.1.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.31.2
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
zope.interface==5.4.0
